fix: code update ad clean up

+++ src/main/java/org/lamisplus/modules/pharmacy/service/DrugOrderService.java
@@ -182,7 +182,7 @@ public class DrugOrderService {
         return patientDrugOrderDTOS;
     }
 
-    public List<DrugOrderDTO> getAllDrugOrdersByVisitId(Long id) {
+    public List<DrugOrderDTO> getAllDrugOrdersByVisitId(Integer id) {
         return drugOrderRepository.findAllByArchivedAndVisitId(UN_ARCHIVED, id)
                 .stream()
                 .map(this::transformDrugOrder)
